+++ aurora-base/article-common/src/main/java/xyz/xcye/article/vo/LinkVO.java
@@ -6,14 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
-import xyz.xcye.core.constant.FieldLengthConstant;
-import xyz.xcye.core.valid.Delete;
-import xyz.xcye.core.valid.Insert;
-import xyz.xcye.core.valid.Update;
-import xyz.xcye.core.valid.validator.ValidateString;
 
-import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Null;
 import java.io.Serializable;
 
@@ -87,7 +80,7 @@ public class LinkVO implements Serializable {
     * 此友情链接对应的站长的qq号
     */
     @ApiModelProperty(value="此友情链接对应的站长的qq号")
-    private Integer qqNumber;
+    private Long qqNumber;
 
     /**
     * 创建时间
